requirements.txt
-i https://pypi.org/simple
arabic-reshaper==3.0.0
asn1crypto==1.5.1
certifi==2023.7.22; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.3; python_version >= '3.7'
cssselect2==0.7.0; python_version >= '3.7'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.4; python_version >= '3.5'
jinja2==3.1.2; python_version >= '3.7'
lxml==4.9.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.3; python_version >= '3.7'
oscrypto==1.3.0
pillow==10.0.0; python_version >= '3.8'
pycparser==2.21
pyhanko==0.20.0; python_version >= '3.7'
pyhanko-certvalidator==0.23.0; python_version >= '3.7'
pypdf==3.16.0; python_version >= '3.6'
pypng==0.20220715.0
python-bidi==0.4.2
python-docx==0.8.11
pyyaml==6.0.1; python_version >= '3.6'
qrcode==7.4.2; python_version >= '3.7'
reportlab==3.6.13; python_version >= '3.7' and python_version < '4'
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
svglib==1.5.1; python_version >= '3.7'
tinycss2==1.2.1; python_version >= '3.7'
typing-extensions==4.7.1; python_version >= '3.7'
tzlocal==5.0.1; python_version >= '3.7'
uritools==4.0.2; python_version >= '3.7'
urllib3==2.0.4; python_version >= '3.7'
webencodings==0.5.1
xhtml2pdf==0.2.11
